Undo changes needed to move the test folder
