+++ .prettierignore
@@ -19,7 +19,6 @@ build-tools/packages/build-cli/README.md
 build-tools/packages/readme-command/README.md
 build-tools/packages/version-tools/README.md
 **/src/**/test/types/*
-**/test/types/*
 **/src/packageVersion.ts
 **/build-tools/CHANGELOG.md
 
@@ -42,7 +41,7 @@ build-tools/packages/build-tools/src/test/data/
 experimental/dds/tree/src/test/documents/
 packages/dds/merge-tree/src/test/literature/*.txt
 packages/dds/merge-tree/src/test/results/*.json
-packages/dds/sequence/test/snapshots/**/*.json
+packages/dds/sequence/src/test/snapshots/**/*.json
 packages/drivers/odsp-driver/src/test/**/*.json
 packages/framework/attributor/src/test/attribution/documents/**/*.json
 packages/test/snapshots/content/
