+++ packages/dds/sequence/src/test/collections.intervalTree.ts
@@ -4,7 +4,7 @@
  */
 
 import { strict as assert } from "assert";
-import { IInterval, IntervalTree } from "../src/intervalTree";
+import { IInterval, IntervalTree } from "../intervalTree";
 
 class TestInterval implements IInterval {
 	constructor(public start: number, public end: number) {}
