+++ packages/dds/sequence/src/test/intervalCollection.fuzz.spec.ts
@@ -14,8 +14,8 @@ import {
 } from "@fluid-internal/stochastic-test-utils";
 import { createDDSFuzzSuite, DDSFuzzModel } from "@fluid-internal/test-dds-utils";
 import { PropertySet } from "@fluidframework/merge-tree";
-import { IntervalCollection, SequenceInterval } from "../src/intervalCollection";
-import { SharedStringFactory } from "../src/sequenceFactory";
+import { IntervalCollection, SequenceInterval } from "../intervalCollection";
+import { SharedStringFactory } from "../sequenceFactory";
 import { assertEquivalentSharedStrings } from "./intervalUtils";
 import {
 	Operation,
