+++ packages/dds/sequence/src/test/intervalCollection.perf.spec.ts
@@ -5,9 +5,9 @@
 
 import { BenchmarkType, benchmark } from "@fluid-tools/benchmark";
 import { MockFluidDataStoreRuntime } from "@fluidframework/test-runtime-utils";
-import { SharedString } from "../src/sharedString";
-import { SharedStringFactory } from "../src/sequenceFactory";
-import { IntervalCollection, IntervalType, SequenceInterval } from "../src/intervalCollection";
+import { SharedString } from "../sharedString";
+import { SharedStringFactory } from "../sequenceFactory";
+import { IntervalCollection, IntervalType, SequenceInterval } from "../intervalCollection";
 
 /**
  * Note: Merge-tree has a number of perf tests for core operations (insert, remove, annotate).
