+++ packages/dds/sequence/src/test/intervalCollection.spec.ts
@@ -14,9 +14,9 @@ import {
 	MockStorage,
 	MockEmptyDeltaConnection,
 } from "@fluidframework/test-runtime-utils";
-import { SharedString } from "../src/sharedString";
-import { SharedStringFactory } from "../src/sequenceFactory";
-import { IntervalCollection, IntervalType, SequenceInterval } from "../src/intervalCollection";
+import { SharedString } from "../sharedString";
+import { SharedStringFactory } from "../sequenceFactory";
+import { IntervalCollection, IntervalType, SequenceInterval } from "../intervalCollection";
 
 const assertIntervals = (
 	sharedString: SharedString,
