+++ packages/dds/sequence/src/test/sequenceDeltaEvent.spec.ts
@@ -15,7 +15,7 @@ import {
 } from "@fluidframework/merge-tree";
 // eslint-disable-next-line import/no-internal-modules
 import { TestClient } from "@fluidframework/merge-tree/dist/test/";
-import { SequenceDeltaEvent } from "../src/sequenceDeltaEvent";
+import { SequenceDeltaEvent } from "../sequenceDeltaEvent";
 
 interface IExpectedSegmentInfo {
 	offset: number;
