+++ packages/dds/sequence/src/test/snapshots/emptyPropsAtEnd.json
