+++ packages/dds/sequence/src/test/snapshots/legacy/headerAndBody.json
