+++ packages/dds/sequence/src/test/snapshots/legacy/headerOnly.json
