+++ packages/dds/sequence/src/test/snapshots/legacy/withIntervals.json
