+++ packages/dds/sequence/src/test/snapshots/legacyWithCatchUp/headerOnly.json
