+++ packages/dds/sequence/src/test/snapshots/legacyWithCatchUp/withAnnotations.json
