+++ packages/dds/sequence/src/test/snapshots/legacyWithCatchUp/withIntervals.json
