+++ packages/dds/sequence/src/test/snapshots/v1/withIntervals.json
