+++ packages/dds/sequence/src/test/snapshots/v1Intervals/withV1Intervals.json
