+++ packages/dds/sequence/src/test/types/validateSequencePrevious.generated.ts
@@ -7,7 +7,7 @@
  * Generated by fluid-type-test-generator in @fluidframework/build-tools.
  */
 import * as old from "@fluidframework/sequence-previous";
-import * as current from "../../src/index";
+import * as current from "../../index";
 
 type TypeOnly<T> = {
     [P in keyof T]: TypeOnly<T[P]>;
